fix namespace import
