+++ ironfish/src/rpc/routes/index.ts
@@ -13,3 +13,4 @@ export * from './miner'
 export * from './faucet'
 export * from './worker'
 export * from './mempool'
+export * from './namespaces'
