+++ ironfish/src/rpc/routes/namespaces.ts
@@ -1,0 +1,17 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+export enum ApiNamespace {
+  chain = 'chain',
+  config = 'config',
+  event = 'event',
+  faucet = 'faucet',
+  miner = 'miner',
+  node = 'node',
+  peer = 'peer',
+  wallet = 'wallet',
+  worker = 'worker',
+  rpc = 'rpc',
+  mempool = 'mempool',
+}
